requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
aiosqlite==0.20.0
asyncpg==0.30.0
greenlet==3.1.1
httpx==0.27.2
python-dotenv==1.0.1
google-api-python-client==2.149.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
jinja2==3.1.4
python-multipart==0.0.12
resend==2.4.0
Pillow>=10.4.0
